Move Markdown into 'utilities' folder.
