+++ resources/assets/components/Block/LegacyContentBlock.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { PuckWaypoint } from '@dosomething/puck-client';
 
-import SectionHeader from '../SectionHeader';
-import Markdown from '../Markdown';
 import { Flex, FlexCell } from '../Flex';
+import SectionHeader from '../SectionHeader';
+import Markdown from '../utilities/Markdown/Markdown';
 
 import './legacy-content-block.scss';
