+++ resources/assets/components/LedeBanner/templates/MosaicTemplate.js
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { Link } from 'react-router-dom';
 
-import Markdown from '../../Markdown';
 import SignupButton from '../../SignupButton';
 import SponsorPromotion from '../../SponsorPromotion';
 import { contentfulImageUrl } from '../../../helpers';
+import Markdown from '../../utilities/Markdown/Markdown';
 import CampaignSignupArrow from '../../CampaignSignupArrow';
 import AffiliateOptionContainer from '../../AffiliateOption';
