+++ resources/assets/components/actions/LinkAction/LinkAction.js
@@ -5,10 +5,10 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
 import Embed from '../../Embed';
-import Markdown from '../../Markdown';
 import Card from '../../utilities/Card/Card';
 import Button from '../../utilities/Button/Button';
 import SponsorPromotion from '../../SponsorPromotion';
+import Markdown from '../../utilities/Markdown/Markdown';
 import { trackPuckEvent } from '../../../helpers/analytics';
 import { isExternal, dynamicString } from '../../../helpers';
