+++ resources/assets/components/actions/ShareAction/ShareAction.js
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Embed from '../../Embed';
-import Markdown from '../../Markdown';
 import Card from '../../utilities/Card/Card';
 import Modal from '../../utilities/Modal/Modal';
 import Button from '../../utilities/Button/Button';
 import ContentfulEntry from '../../ContentfulEntry';
+import Markdown from '../../utilities/Markdown/Markdown';
 import { trackPuckEvent } from '../../../helpers/analytics';
 import {
   dynamicString,
