+++ resources/assets/components/utilities/Markdown/markdown.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox.scss';
+@import '../../../scss/next-toolbox.scss';
 
 .markdown {
   h1,
